requirements.txt
django==1.7.3

djangorestframework==3.0.4
